+++ CHANGELOG.md
@@ -20,6 +20,8 @@ You can find our backwards-compatibility policy [here](https://github.com/hynek/
 - `structlog.processors.LogfmtRenderer` now escapes newlines.
   [#592](https://github.com/hynek/structlog/pull/592)
 
+- `structlog.processors.LogfmtRenderer` now escapes backslashes.
+
 
 ## [24.1.0](https://github.com/hynek/structlog/compare/23.3.0...24.1.0) - 2024-01-08